% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{calculate_luminance}
\alias{calculate_luminance}
\title{Calculate relative luminance of a color}
\usage{
calculate_luminance(hex)
}
\arguments{
\item{hex}{Hex color code}
}
\value{
Numeric value between 0 and 1 representing relative luminance
}
\description{
Calculate relative luminance of a color
}
\keyword{internal}
