% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{clean_pool}
\alias{clean_pool}
\title{Clean Database Pool}
\usage{
clean_pool(session)
}
\arguments{
\item{session}{Shiny session object}
}
\description{
Closes the database connection pool and performs cleanup operations
when the application is shutting down.
}
