% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cleanup_app}
\alias{cleanup_app}
\title{Clean up application resources}
\usage{
cleanup_app(session)
}
\arguments{
\item{session}{Shiny session object}
}
\description{
Closes the database connection pool and performs cleanup operations
when the application is shutting down.
}
