% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{cleanup_pool}
\alias{cleanup_pool}
\title{Clean Up Database Pool}
\usage{
cleanup_pool(pool)
}
\arguments{
\item{pool}{Database pool object to close}
}
\description{
Closes the database pool and removes lock files.
}
