% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_ops}
\alias{db_ops}
\title{Database Operations Class}
\description{
R6 Class for managing database operations related to survey data storage
and retrieval using PostgreSQL.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{session_id}}{Character. Unique identifier for the current session}

\item{\code{pool}}{Pool object. Database connection pool}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-db_ops-new}{\code{db_ops$new()}}
\item \href{#method-db_ops-create_survey_data_table}{\code{db_ops$create_survey_data_table()}}
\item \href{#method-db_ops-check_table_exists}{\code{db_ops$check_table_exists()}}
\item \href{#method-db_ops-execute_db_operation}{\code{db_ops$execute_db_operation()}}
\item \href{#method-db_ops-update_survey_data_table}{\code{db_ops$update_survey_data_table()}}
\item \href{#method-db_ops-clone}{\code{db_ops$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_ops-new"></a>}}
\if{latex}{\out{\hypertarget{method-db_ops-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new db_ops instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{db_ops$new(pool, session_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pool}}{Database connection pool}

\item{\code{session_id}}{Session identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new db_ops object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_ops-create_survey_data_table"></a>}}
\if{latex}{\out{\hypertarget{method-db_ops-create_survey_data_table}{}}}
\subsection{Method \code{create_survey_data_table()}}{
Create a new survey data table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{db_ops$create_survey_data_table(survey_name, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{survey_name}}{Character. Name of the survey}

\item{\code{data}}{data.frame. Survey data to be stored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character. Name of the created table (invisible)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_ops-check_table_exists"></a>}}
\if{latex}{\out{\hypertarget{method-db_ops-check_table_exists}{}}}
\subsection{Method \code{check_table_exists()}}{
Check if a table exists in the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{db_ops$check_table_exists(table_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{Character. Name of the table to check}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical indicating if table exists
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_ops-execute_db_operation"></a>}}
\if{latex}{\out{\hypertarget{method-db_ops-execute_db_operation}{}}}
\subsection{Method \code{execute_db_operation()}}{
Execute a database operation with appropriate connection handling
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{db_ops$execute_db_operation(operation, error_message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{operation}}{Function. The database operation to execute}

\item{\code{error_message}}{Character. Error message if operation fails}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Result of the database operation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_ops-update_survey_data_table"></a>}}
\if{latex}{\out{\hypertarget{method-db_ops-update_survey_data_table}{}}}
\subsection{Method \code{update_survey_data_table()}}{
Update existing survey data table with new data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{db_ops$update_survey_data_table(survey_name, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{survey_name}}{Character. Name of the survey}

\item{\code{data}}{data.frame. New survey data to append}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character. Name of the updated table (invisible)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_ops-clone"></a>}}
\if{latex}{\out{\hypertarget{method-db_ops-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{db_ops$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
