% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_config.R
\name{format_choices_for_js}
\alias{format_choices_for_js}
\title{Format Choices for JavaScript Survey Library}
\usage{
format_choices_for_js(
  choices,
  is_parent = FALSE,
  child_field = NULL,
  is_child = FALSE,
  parent_field = NULL,
  display_col = NULL
)
}
\arguments{
\item{choices}{Either a vector of choices or a list of hierarchical choices}

\item{is_parent}{Logical, whether this field is a parent field}

\item{child_field}{Character, name of the field containing child choices}

\item{is_child}{Logical, whether this field contains child choices}

\item{parent_field}{Character, name of the field containing parent choices}

\item{display_col}{Character, optional name of column containing display text}
}
\value{
A list formatted for JavaScript survey components with relationship metadata
}
\description{
Formats R choice data structures into a format compatible with JavaScript survey components.
Handles both flat lists of choices and hierarchical choice structures, with support for
parent-child relationships between fields.
}
\examples{
\dontrun{
# Parent choices
choices <- c("parent1", "parent2")
formatted <- format_choices_for_js(choices, is_parent = TRUE, child_field = "child_field")

# Child choices
child_choices <- list(
  list(value = "child1", text = "Child 1", parentId = 1),
  list(value = "child2", text = "Child 2", parentId = 1)
)
formatted_children <- format_choices_for_js(child_choices, is_child = TRUE, parent_field = "parent_field")
}
}
