% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{get_contrast_color}
\alias{get_contrast_color}
\title{Get contrasting text color}
\usage{
get_contrast_color(background_hex)
}
\arguments{
\item{background_hex}{Background hex color code}
}
\value{
Hex color code for text ("#000000" for dark text or "#ffffff" for light text)
}
\description{
Get contrasting text color
}
\keyword{internal}
