% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_config.R
\name{get_unique_field_values}
\alias{get_unique_field_values}
\title{Get Existing Values for Unique Fields}
\usage{
get_unique_field_values(dynamic_config, db_ops, write_table)
}
\arguments{
\item{dynamic_config}{List of configuration entries}

\item{db_ops}{Database operations object that contains methods for reading tables}

\item{write_table}{Table name to check}
}
\value{
List of unique field values by field name
}
\description{
Retrieves existing values from the database for fields that require uniqueness validation
}
