% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{hide_and_show}
\alias{hide_and_show}
\title{Switch Visibility Between Two Elements}
\usage{
hide_and_show(hide_id, show_id)
}
\arguments{
\item{hide_id}{String. ID of the element to hide.}

\item{show_id}{String. ID of the element to show.}
}
\description{
Shows one element while hiding another element via shinyjs. Provides a simple
toggle between two HTML elements without animations.
}
\keyword{internal}
