% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{hide_and_show_message}
\alias{hide_and_show_message}
\title{Helper function to hide one message and show another}
\usage{
hide_and_show_message(hide_id, show_id, fade_time = 1)
}
\arguments{
\item{hide_id}{ID of the message to hide}

\item{show_id}{ID of the message to show}

\item{fade_time}{Time in seconds for fade animation (default: 1)}
}
\description{
Helper function to hide one message and show another
}
