% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{is_light_color}
\alias{is_light_color}
\title{Determine if a color is light or dark}
\usage{
is_light_color(hex, threshold = 0.35)
}
\arguments{
\item{hex}{Hex color code}

\item{threshold}{Luminance threshold (default: 0.179, based on WCAG guidelines)}
}
\value{
Boolean indicating if color is light (TRUE) or dark (FALSE)
}
\description{
Determine if a color is light or dark
}
\keyword{internal}
