% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_ldap.R
\name{ldap_login_server}
\alias{ldap_login_server}
\title{LDAP login server module}
\usage{
ldap_login_server(id, ldap_auth, logger = NULL)
}
\arguments{
\item{id}{Module ID}

\item{ldap_auth}{LdapAuth instance}

\item{logger}{Optional logger}
}
\description{
LDAP login server module
}
