% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_ldap.R
\name{ldap_login_ui}
\alias{ldap_login_ui}
\title{Create LDAP login UI}
\usage{
ldap_login_ui(id, title = "Login", logo = NULL, theme_color = "#007bff")
}
\arguments{
\item{id}{Namespace ID}

\item{title}{Login form title}

\item{logo}{Logo URL to display instead of title (optional)}

\item{theme_color}{Hex color code for button styling (optional)}
}
\description{
Create LDAP login UI
}
