% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{hex_to_rgb}
\alias{hex_to_rgb}
\title{Convert hex color to RGB values}
\usage{
hex_to_rgb(hex)
}
\arguments{
\item{hex}{Hex color code}
}
\value{
Numeric vector of RGB values
}
\description{
Convert hex color to RGB values
}
\keyword{internal}
