% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hide_and_show}
\alias{hide_and_show}
\title{Switch Visibility Between Two Elements}
\usage{
hide_and_show(hide_id, show_id)
}
\arguments{
\item{hide_id}{Character string specifying the ID of the element to hide}

\item{show_id}{Character string specifying the ID of the element to show}
}
\description{
Shows one element while hiding another. Simple toggle between two DIV elements
without animations.
}
\keyword{internal}
