% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{survey_single}
\alias{survey_single}
\title{Create a single survey}
\usage{
survey_single(
  json,
  show_responses = TRUE,
  theme = "defaultV2",
  theme_primary = "#003594",
  theme_mode = "light"
)
}
\arguments{
\item{json}{Survey JSON string or object}

\item{show_responses}{Whether to show responses after submission}

\item{theme}{Theme name ("defaultV2" or "modern")}

\item{primary}{Primary color hex colde (optional)}
}
\description{
Create a single survey
}
